refactor create and update rountine
