+++ src/modules/routine/application/interface/routine.interface.ts
@@ -1,0 +1,12 @@
+export interface IRoutine {
+  name: string;
+  description: string;
+  exercises: {
+    exerciseId: string;
+    sets: number;
+    reps: number;
+    rest: number;
+  }[];
+
+  createdBy: string;
+}
