--- src/modules/routine/application/interface/routineinterface.ts
@@ -1,9 +1,0 @@
-export interface IRoutine {
-  name: string;
-  description: string;
-  exercises: string[];
-  sets: number;
-  reps: number;
-  rest: number;
-  createdBy: string;
-}
